+++ src/components/Header.jsx
@@ -1,4 +1,4 @@
-import Dropdown from "../components/Dropdown";
+import Dropdown from "./Dropdown";
 
 function Header() {
   return (
