+++ src/pages/Signup.jsx
@@ -107,12 +107,13 @@ function Signup() {
               <div className="pt-5 mt-6 border-t border-slate-200">
                 <div className="text-sm">
                   Have an account?
-                  <p
+                  <a
                     className="font-medium text-indigo-500 hover:text-indigo-600 cursor-pointer"
-                    to="/signin"
+                    href="/signin"
                   >
+                    {" "}
                     Sign In
-                  </p>
+                  </a>
                 </div>
               </div>
             </div>
